+++ src/libs/Detach/Inline.cs
@@ -11,12 +11,11 @@ namespace Detach;
 /// </summary>
 public static partial class Inline
 {
-	private const string _separatorUtf16 = ", ";
-
 	private static readonly byte[] _bufferUtf8 = new byte[2048];
 	private static readonly char[] _bufferUtf16 = new char[2048];
 
-	private static ReadOnlySpan<byte> SeparatorUtf8 => ", "u8;
+	internal static ReadOnlySpan<byte> NumericSeparatorUtf8 => ", "u8;
+	internal static ReadOnlySpan<char> NumericSeparatorUtf16 => ", ";
 
 	public static Span<byte> BufferUtf8 => _bufferUtf8;
 	public static Span<char> BufferUtf16 => _bufferUtf16;
@@ -73,6 +72,14 @@ public static partial class Inline
 	private static void WriteUtf16(ref int charsWritten, string value)
 	{
 		if (!value.AsSpan().TryCopyTo(BufferUtf16[charsWritten..]))
+			throw new InvalidOperationException("The formatted string is too long.");
+
+		charsWritten += value.Length;
+	}
+
+	private static void WriteUtf16(ref int charsWritten, ReadOnlySpan<char> value)
+	{
+		if (!value.TryCopyTo(BufferUtf16[charsWritten..]))
 			throw new InvalidOperationException("The formatted string is too long.");
 
 		charsWritten += value.Length;